Enlarged "maximum" integer values. SCTP work.
git-svn-id: https://svn.code.sf.net/p/spot-on/code/branches/0.09.01@2651 0a995953-517f-4c62-b348-104708d926b2

+++ Common/spot-on-common.h
@@ -46,15 +46,15 @@ namespace spoton_common
 						 ** will be represented in
 						 ** base sixteen.
 						 */
-  static const int MAXIMUM_NEIGHBOR_BUFFER_SIZE =
+  static const qint64 MAXIMUM_NEIGHBOR_BUFFER_SIZE =
     10485760; /*
 	      ** The buffer size
 	      ** should be larger
 	      ** than the content
 	      ** length.
 	      */
-  static const int MAXIMUM_NEIGHBOR_CONTENT_LENGTH = 5242880;
-  static const int MINIMUM_NEIGHBOR_CONTENT_LENGTH = 256;
+  static const qint64 MAXIMUM_NEIGHBOR_CONTENT_LENGTH = 5242880;
+  static const qint64 MINIMUM_NEIGHBOR_CONTENT_LENGTH = 256;
   static const int NAME_MAXIMUM_LENGTH = 16;
 }
